+++ app/src/routeTree.gen.ts
@@ -23,6 +23,7 @@ import { Route as AuthedAdminBoundariesRouteImport } from './routes/_authed/admi
 import { Route as AuthedAdminComputersRouteImport } from './routes/_authed/admin/computers'
 import { Route as AuthedAdminConnectorsRouteImport } from './routes/_authed/admin/connectors'
 import { Route as AuthedAdminCredentialsRouteImport } from './routes/_authed/admin/credentials'
+import { Route as AuthedAdminIdentityProvidersRouteImport } from './routes/_authed/admin/identity-providers'
 import { Route as AuthedAdminPeopleRouteImport } from './routes/_authed/admin/people'
 import { Route as AuthedAdminPlaygroundRouteImport } from './routes/_authed/admin/playground'
 import { Route as AuthedAdminPluginsRouteImport } from './routes/_authed/admin/plugins'
@@ -104,6 +105,12 @@ const AuthedAdminCredentialsRoute = AuthedAdminCredentialsRouteImport.update({
   path: '/credentials',
   getParentRoute: () => AuthedAdminRouteRoute,
 } as any)
+const AuthedAdminIdentityProvidersRoute =
+  AuthedAdminIdentityProvidersRouteImport.update({
+    id: '/identity-providers',
+    path: '/identity-providers',
+    getParentRoute: () => AuthedAdminRouteRoute,
+  } as any)
 const AuthedAdminPeopleRoute = AuthedAdminPeopleRouteImport.update({
   id: '/people',
   path: '/people',
@@ -183,6 +190,7 @@ export interface FileRoutesByFullPath {
   '/admin/computers': typeof AuthedAdminComputersRoute
   '/admin/connectors': typeof AuthedAdminConnectorsRouteWithChildren
   '/admin/credentials': typeof AuthedAdminCredentialsRoute
+  '/admin/identity-providers': typeof AuthedAdminIdentityProvidersRoute
   '/admin/people': typeof AuthedAdminPeopleRoute
   '/admin/playground': typeof AuthedAdminPlaygroundRoute
   '/admin/plugins': typeof AuthedAdminPluginsRoute
@@ -207,6 +215,7 @@ export interface FileRoutesByTo {
   '/admin/computers': typeof AuthedAdminComputersRoute
   '/admin/connectors': typeof AuthedAdminConnectorsRouteWithChildren
   '/admin/credentials': typeof AuthedAdminCredentialsRoute
+  '/admin/identity-providers': typeof AuthedAdminIdentityProvidersRoute
   '/admin/people': typeof AuthedAdminPeopleRoute
   '/admin/playground': typeof AuthedAdminPlaygroundRoute
   '/admin/plugins': typeof AuthedAdminPluginsRoute
@@ -235,6 +244,7 @@ export interface FileRoutesById {
   '/_authed/admin/computers': typeof AuthedAdminComputersRoute
   '/_authed/admin/connectors': typeof AuthedAdminConnectorsRouteWithChildren
   '/_authed/admin/credentials': typeof AuthedAdminCredentialsRoute
+  '/_authed/admin/identity-providers': typeof AuthedAdminIdentityProvidersRoute
   '/_authed/admin/people': typeof AuthedAdminPeopleRoute
   '/_authed/admin/playground': typeof AuthedAdminPlaygroundRoute
   '/_authed/admin/plugins': typeof AuthedAdminPluginsRoute
@@ -264,6 +274,7 @@ export interface FileRouteTypes {
     | '/admin/computers'
     | '/admin/connectors'
     | '/admin/credentials'
+    | '/admin/identity-providers'
     | '/admin/people'
     | '/admin/playground'
     | '/admin/plugins'
@@ -288,6 +299,7 @@ export interface FileRouteTypes {
     | '/admin/computers'
     | '/admin/connectors'
     | '/admin/credentials'
+    | '/admin/identity-providers'
     | '/admin/people'
     | '/admin/playground'
     | '/admin/plugins'
@@ -315,6 +327,7 @@ export interface FileRouteTypes {
     | '/_authed/admin/computers'
     | '/_authed/admin/connectors'
     | '/_authed/admin/credentials'
+    | '/_authed/admin/identity-providers'
     | '/_authed/admin/people'
     | '/_authed/admin/playground'
     | '/_authed/admin/plugins'
@@ -436,6 +449,13 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof AuthedAdminCredentialsRouteImport
       parentRoute: typeof AuthedAdminRouteRoute
     }
+    '/_authed/admin/identity-providers': {
+      id: '/_authed/admin/identity-providers'
+      path: '/identity-providers'
+      fullPath: '/admin/identity-providers'
+      preLoaderRoute: typeof AuthedAdminIdentityProvidersRouteImport
+      parentRoute: typeof AuthedAdminRouteRoute
+    }
     '/_authed/admin/people': {
       id: '/_authed/admin/people'
       path: '/people'
@@ -542,6 +562,7 @@ interface AuthedAdminRouteRouteChildren {
   AuthedAdminComputersRoute: typeof AuthedAdminComputersRoute
   AuthedAdminConnectorsRoute: typeof AuthedAdminConnectorsRouteWithChildren
   AuthedAdminCredentialsRoute: typeof AuthedAdminCredentialsRoute
+  AuthedAdminIdentityProvidersRoute: typeof AuthedAdminIdentityProvidersRoute
   AuthedAdminPeopleRoute: typeof AuthedAdminPeopleRoute
   AuthedAdminPlaygroundRoute: typeof AuthedAdminPlaygroundRoute
   AuthedAdminPluginsRoute: typeof AuthedAdminPluginsRoute
@@ -556,6 +577,7 @@ const AuthedAdminRouteRouteChildren: AuthedAdminRouteRouteChildren = {
   AuthedAdminComputersRoute: AuthedAdminComputersRoute,
   AuthedAdminConnectorsRoute: AuthedAdminConnectorsRouteWithChildren,
   AuthedAdminCredentialsRoute: AuthedAdminCredentialsRoute,
+  AuthedAdminIdentityProvidersRoute: AuthedAdminIdentityProvidersRoute,
   AuthedAdminPeopleRoute: AuthedAdminPeopleRoute,
   AuthedAdminPlaygroundRoute: AuthedAdminPlaygroundRoute,
   AuthedAdminPluginsRoute: AuthedAdminPluginsRoute,